requirements.txt
streamlit
requests
beautifulsoup4
python-whois
urllib3
tldextract